⚡ 배송 상태 변경, Chucks 아이템 처리자로 데이터 받아오기

See merge request s08-final/S08P31B306!200

+++ backend/reslow/src/main/java/com/ssafy/reslow/domain/order/dto/OrderRegistRequest.java
@@ -8,7 +8,7 @@ public class OrderRegistRequest {
 	private Long productNo;
 	private Long issuedCouponNo;
 	private String recipient;
-	private Long zipcode;
+	private int zipcode;
 	private String address;
 	private String addressDetail;
 	private String phoneNumber;
